Complete refactor of code

+++ Setup/Uninstall.php
@@ -2,41 +2,27 @@
 
 namespace Apsis\One\Setup;
 
-use Apsis\One\Model\Service\Log as ApsisLogHelper;
-use Magento\Authorization\Model\Acl\Role\Group as RoleGroup;
-use Magento\Authorization\Model\UserContextInterface;
 use Magento\Framework\Setup\ModuleContextInterface;
 use Magento\Framework\Setup\SchemaSetupInterface;
 use Magento\Framework\Setup\SetupInterface;
 use Magento\Framework\Setup\UninstallInterface;
-use Apsis\One\Model\Service\Core as ApsisCoreHelper;
+use Apsis\One\Service\BaseService;
 use Throwable;
 
 class Uninstall implements UninstallInterface
 {
     /**
-     * @var ApsisLogHelper
+     * @var BaseService
      */
-    private ApsisLogHelper $logHelper;
+    private BaseService $baseService;
 
     /**
-     * Uninstall constructor.
-     *
-     * @param ApsisLogHelper $logHelper
+     * @param BaseService $baseService
      */
-    public function __construct(ApsisLogHelper $logHelper)
+    public function __construct(BaseService $baseService)
     {
-        $this->logHelper = $logHelper;
+        $this->baseService = $baseService;
     }
-
-    /**
-     * @var array
-     */
-    protected array $apsisTablesArr = [
-        ApsisCoreHelper::APSIS_ABANDONED_TABLE,
-        ApsisCoreHelper::APSIS_EVENT_TABLE,
-        ApsisCoreHelper::APSIS_PROFILE_TABLE
-    ];
 
     /**
      * @param SchemaSetupInterface $setup
@@ -44,10 +30,10 @@ class Uninstall implements UninstallInterface
      *
      * @return void
      */
-    public function uninstall(SchemaSetupInterface $setup, ModuleContextInterface $context)
+    public function uninstall(SchemaSetupInterface $setup, ModuleContextInterface $context): void
     {
         try {
-            $this->logHelper->log(__METHOD__);
+            $this->baseService->log(__METHOD__);
 
             //Remove all module tables
             $this->removeAllModuleTables($setup);
@@ -55,7 +41,7 @@ class Uninstall implements UninstallInterface
             // Remove all module data from Magento tables
             $this->removeAllModuleDataFromMagentoTables($setup);
         } catch (Throwable $e) {
-            $this->logHelper->logError(__METHOD__, $e);
+            $this->baseService->logError(__METHOD__, $e);
         }
     }
 
@@ -66,10 +52,16 @@ class Uninstall implements UninstallInterface
      */
     public function removeAllModuleTables(SetupInterface $setup): void
     {
-        $this->logHelper->log(__METHOD__);
+        try {
+            $this->baseService->log(__METHOD__);
 
-        foreach ($this->apsisTablesArr as $tableName) {
-            $setup->getConnection()->dropTable($setup->getTable($tableName));
+            foreach (array_merge(array_keys(InstallSchema::TABLES), ['apsis_profile_batch']) as $tableName) {
+                if ($setup->getConnection()->isTableExists($setup->getTable($tableName))) {
+                    $setup->getConnection()->dropTable($setup->getTable($tableName));
+                }
+            }
+        } catch (Throwable $e) {
+            $this->baseService->logError(__METHOD__, $e);
         }
     }
 
@@ -80,45 +72,49 @@ class Uninstall implements UninstallInterface
      */
     public function removeAllModuleDataFromMagentoTables(SetupInterface $setup): void
     {
-        $this->logHelper->log(__METHOD__);
+        try {
+            $this->baseService->log(__METHOD__);
 
-        //Remove all module config
-        $setup->getConnection()->delete(
-            $setup->getTable('core_config_data'),
-            "path like 'apsis_one%'"
-        );
-
-        //Remove rules belonging to APSIS role
-        $select = $setup->getConnection()->select()
-            ->from($setup->getTable('authorization_role'), 'role_id')
-            ->where('role_name = ?', 'APSIS Support Agent');
-        $role = $setup->getConnection()->fetchOne($select);
-        if ($role) {
+            //Remove all module config
             $setup->getConnection()->delete(
-                $setup->getTable('authorization_rule'),
-                ['role_id = ?' => (int) $role]
+                $setup->getTable('core_config_data'),
+                "path like 'apsis_one%'"
             );
+
+            //Remove rules belonging to APSIS role
+            $select = $setup->getConnection()->select()
+                ->from($setup->getTable('authorization_role'), 'role_id')
+                ->where('role_name = ?', 'APSIS Support Agent');
+            $role = $setup->getConnection()->fetchOne($select);
+            if ($role) {
+                $setup->getConnection()->delete(
+                    $setup->getTable('authorization_rule'),
+                    ['role_id = ?' => (int) $role]
+                );
+            }
+
+            //Remove role created by the module
+            $setup->getConnection()->delete(
+                $setup->getTable('authorization_role'),
+                ['role_name = ?' => 'APSIS Support Agent']
+            );
+
+            //Remove all ui bookmarks belonging to module to force rebuild new ui bookmarks
+            $setup->getConnection()->delete(
+                $setup->getTable('ui_bookmark'),
+                $setup->getConnection()->quoteInto(
+                    'namespace in (?)',
+                    ['apsis_abandoned_grid', 'apsis_event_grid', 'apsis_profile_grid']
+                )
+            );
+
+            //Removed all cron jobs.
+            $setup->getConnection()->delete(
+                $setup->getTable('cron_schedule'),
+                "job_code like 'apsis_one%'"
+            );
+        } catch (Throwable $e) {
+            $this->baseService->logError(__METHOD__, $e);
         }
-
-        //Remove role created by the module
-        $setup->getConnection()->delete(
-            $setup->getTable('authorization_role'),
-            ['role_name = ?' => 'APSIS Support Agent']
-        );
-
-        //Remove all ui bookmarks belonging to module to force rebuild new ui bookmarks
-        $setup->getConnection()->delete(
-            $setup->getTable('ui_bookmark'),
-            $setup->getConnection()->quoteInto(
-                'namespace in (?)',
-                ['apsis_abandoned_grid', 'apsis_event_grid', 'apsis_profile_grid']
-            )
-        );
-
-        //Removed all cron jobs.
-        $setup->getConnection()->delete(
-            $setup->getTable('cron_schedule'),
-            "job_code like 'apsis_one%'"
-        );
     }
 }
